Update DOI handling and add warning for jhu deposit agreement
