+++ app/templates/components/workflow-metadata.hbs
@@ -2,5 +2,5 @@
 <h4 class="font-weight-light pull-right" style="margin-top:6.5px;">
   Form {{displayFormStep}} of {{schemas.length}}
 </h4>
-{{metadata-form schema=schema model=model.newSubmission doiInfo=doiInfo currentFormStep=currentFormStep nextForm=(action "nextForm") previousForm=(action "previousForm")}}
+{{metadata-form didNotAgree=didNotAgree schema=schema model=model.newSubmission doiInfo=doiInfo currentFormStep=currentFormStep nextForm=(action "nextForm") previousForm=(action "previousForm")}}
 {{yield}}
